Fix cyclic-import issues
