+++ bluepysnap/circuit_validation.py
@@ -12,46 +12,13 @@ import pandas as pd
 
 from bluepysnap import schemas
 from bluepysnap.config import Parser
+from bluepysnap.exceptions import BluepySnapValidationError as Error
 from bluepysnap.morph import EXTENSIONS_MAPPING
 from bluepysnap.sonata_constants import DEFAULT_EDGE_TYPE, DEFAULT_NODE_TYPE
 from bluepysnap.utils import load_json
 
 L = logging.getLogger("brainbuilder")
 MAX_MISSING_FILES_DISPLAY = 10
-
-
-class Error:
-    """Error used for reporting of validation errors."""
-
-    FATAL = "FATAL"
-    WARNING = "WARNING"
-    INFO = "INFO"
-
-    def __init__(self, level, message=None):
-        """Error.
-
-        Args:
-            level (str): error level
-            message (str|None): message
-        """
-        self.level = level
-        self.message = message
-
-    def __str__(self):
-        """Returns only message by default."""
-        return str(self.message)
-
-    __repr__ = __str__
-
-    def __eq__(self, other):
-        """Two errors are equal if inherit from Error and their level, message are equal."""
-        if not isinstance(other, Error):
-            return False
-        return self.level == other.level and self.message == other.message
-
-    def __hash__(self):
-        """Hash. Errors with the same level and message give the same hash."""
-        return hash(self.level) ^ hash(self.message)
 
 
 def fatal(message):
