+++ bluepysnap/edges/edges.py
@@ -21,7 +21,7 @@ import numpy as np
 
 from bluepysnap._doctools import AbstractDocSubstitutionMeta
 from bluepysnap.circuit_ids import CircuitEdgeIds, CircuitNodeId, CircuitNodeIds
-from bluepysnap.edges import EdgePopulation
+from bluepysnap.edges.edge_population import EdgePopulation
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.network import NetworkObject
