+++ tox.ini
@@ -28,7 +28,7 @@ deps =
     isort
     pycodestyle
     pydocstyle
-    pylint<3.0.0  # overzealous cyclic-import in pylint>=3.0.0, see: https://github.com/pylint-dev/pylint/issues/9124
+    pylint
 commands =
     black --check .
     isort --check {[base]name} tests setup.py doc/source/conf.py
